Prepend RangeWithStepTime to the Range object
after doing some researching on how to actually create individual dates
from a range of dates i stumbeld upon a module that was almost accepted
as part of ruby on rails's Range object.

it provides the exact same behaviour i need and is safe to use because
it defers to the default Range behaviour when not dealing with date/time
i tested this in the console and it just works and is preformant. i can
build on top of this.

prepending it for in an initializer will do for now.

References:
https://github.com/rails/rails/pull/21192/commits/4bedecd4d5f46c15aa50fd532cae69dba4ce3186
https://stackoverflow.com/questions/19093487/ruby-create-range-of-dates#answer-19094504

TODO:
grab the tests from the original rails pull request and create some serious rspec specs for it.