Add rate indicator service
This PORO class basically calculates the rate and the averige rate.

the most important part is the private method:

  def daily_prices
    (begin_date...end_date).step(1.day).map do |date|
      within_seasonal_range?(date) ? listing.seasonal_rate.daily_rate : listing.default_daily_rate
    end
  end

it puts each individual date of the given date range in an array and
checks if it is within_seasonal_range?

if it is... it sets the seasonal_rate price else.. it sets the
defaule_daily_rate as price.

the private methods rate sums up all values in that array.
the private method daily_avg_rate calculates the averige.

invoking any of the 2 main methods hit the database with only 1 query
and it takes about 0.4ms (0.0004s) which is quite preformant.